+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
   - WPM Decay. Used to sword words on slowest WPM.
 - Words page. Shows all the practice words, their WPM and last practiced time.
   - You can delete words from the list.
+  - You can sort by the different columns.
 
 ## 2024-07-06 - 2024-07-31